FIX: Moving component Loader folder
